Adds explicit solr version dependency instructions.
I was unable to get the project working following the instruction in the
README telling me to install `solr` through `brew`. When I explicitly
tried an older version it did work, so I'm working under the assumption
that the stable version has been incremented, and the README has not yet
been updated accordingly.

Note that solr 4 and 5 currently do not work with the instructions listed in the README.
Consider upgrading to solr 4, which seems to be the current stable version.